requirements.txt
yamllint
pyyaml
requests
argparse
geopy>1.0
